The first change in this diff is to the player.rs file, on line 22. The original code had 1 as the value for the player's sprite, but the new code has 3 as the value. This change will make the player's sprite appear as a 3 instead of a 1 when the game is played.
The second change is on line 35, where the code for the player's sprite sheet bundle is changed. In the original code, the sprite's transform was set to scale up the size of the sprite by a factor of 6. However, in the new code, the sprite's transform is set to scale up the size of the sprite by a factor of 3. This change will make the player's sprite appear smaller in the game.

The third and final change is on line 61, where the code for the player's animation is changed. In the original code, the animation was set to repeat every 0.1 seconds. However, in the new code, the animation is set to repeat every 0.5 seconds. This change will make the player's sprite appear to move more slowly in the game.