The first file that is being diffed is the player_sprites.png file. This file contains the sprites for the player character. The second file is the player.rs file. This file contains the Rust code for the player character.
The diff for the player_sprites.png file shows that the file has been changed. The diff for the player.rs file shows that the code for the player character has been changed. The code changes include adding constants for the base x and y positions of the player character, and changing the animation timer code so that the character sprite changes when the timer expires.