This diff contains changes to the src/main.rs and src/player/component.rs files. In src/main.rs, a new module, player, is added and in src/player/component.rs, a new component, PlayerSpriteIndices, is added.
In src/main.rs, the PlayerSpriteIndices component is used to store the indices of the sprites in the player animation. The animation timer is used to control the animation of the player sprite.

In src/player/component.rs, the PlayerAnimationTimer is used to control the animation of the player sprite. The sprite sheet is used to store the texture atlas for the player sprite.