The first diff is for the file src/player/component.rs. In this file, some changes were made to the PlayerAnimationTimer struct. First, a new function called new() was added. This function initializes a new PlayerAnimationTimer. Next, the tick() function was updated to take in a delta time parameter. This allows the animation timer to be more accurate. Finally, the just_finished() function was updated to return true if the timer has just finished.
The second diff is for the file src/player/mod.rs. In this file, some changes were made to the PlayerPlugin struct. First, the new animation module was added. This module contains the code for animating the player. Next, the build() function was updated to add a new system for animating the player. This system uses the new animation module to animate the player.