The diff shows the changes made to the main.rs and player/component.rs files. In main.rs, the camera position and size are changed, and in player/component.rs, the Player struct is annotated with the Reflect component.
The camera position is changed from 400,0 to 500,100 and the size is changed from 256,256 to 1200,800. This will make the camera display a larger area of the game world.

The Player struct is now annotated with the Reflect component. This will allow the struct to be reflected by the game engine, which will make it easier to work with in code.