The first thing to note is that the diff is for multiple files. The main file being changed is src/main.rs, but there are also changes in src/player/component.rs and src/tile.rs.
In the main file, the changes seem to be related to adding a new image plugin to prevent blurry sprites, as well as some changes to the camera setup. In the player file, it looks like the player is being moved, and in the tile file, it looks like the tiles are being translated.