The first change in this diff is to the image plugin. The image plugin is responsible for displaying images in the game. The change to the image plugin will improve the quality of the images displayed in the game.
The second change is to the window plugin. The window plugin is responsible for creating and managing windows in the game. The change to the window plugin will improve the quality of the windows displayed in the game.

The third change is to the camera setup. The camera setup is responsible for setting up the camera in the game. The change to the camera setup will improve the quality of the camera in the game.