The diff contains two parts: one for the file src/game_state/splash.rs and one for the file src/main.rs.
In the file src/game_state/splash.rs, there are changes to the splash_setup function. The function now takes an additional argument, asset_server, and loads an image from the server. The function also now uses the OnExit event to trigger the splash_setup_2 function.

In the file src/main.rs, there are changes to the main function. The function now uses the OnExit event to trigger the game_state::GameState::Menu function.