The first thing to note is that the diff is for multiple files. The main file being changed is src/game_state/splash.rs, but there are also changes to other files in the same directory.
The diff shows that the splash screen plugin is being changed to use a new SplashScreen struct. This struct has a start() function which sets up the splash screen image and a countdown() function which handles the timer for the splash screen. The despawn() function is called when the splash screen is finished.

In the src/game_state/splash.rs file, the changes are mostly just renaming the functions to match the new SplashScreen struct. There are also some changes to the size and position of the splash screen image.