The diff contains changes to the src/game_state/menu.rs, src/game_state/mod.rs, and src/game_state/splash.rs files.
In the src/game_state/menu.rs file, the changes seem to be related to adding a new system for the menu screen. In the src/game_state/mod.rs file, the changes seem to be related to adding a new enum for the menu state. In the src/game_state/splash.rs file, the changes seem to be related to changing the game state from MenuScreen to Menu.