The diff contains changes to three files: README.md, src/main.rs, and src/startup/menu.rs.
In the README.md file, a new line has been added to the end of the file. This new line indicates that the sprite walking animation has been fixed with a timer animation, and that a new line indicating that a new and continue functionality needs to be created has been added.

In the src/main.rs file, a new line has been added to the top of the file, importing the story module. Additionally, in the main function, a new line has been added which adds the StoryState state to the application.

In the src/startup/menu.rs file, four new lines have been added. The first new line adds a new system which views the menu screen. The second new line adds a new system which enters the game when the return key is pressed. The third new line adds a new system which cleans up the menu screen when the player exits the menu. The fourth new line adds a new system which sets the story state to NewGame when the menu is exited.