The first diff is for the file src/main.rs. In this file, some lines have been added and some have been removed. The lines that have been added are 25-26 and 29-30. The lines that have been removed are 21-22 and 27-28.
The second diff is for the file src/miscellanous/textbox.rs. In this file, some lines have been added and some have been removed. The lines that have been added are 33-36. The lines that have been removed are 21-22.

The third diff is for the file src/story/new_game.rs. In this file, some lines have been added and some have been removed. The lines that have been added are 8-9, 23-24, and 26-27. The lines that have been removed are 17-18.