The first thing to note is that the main.rs file has been changed. In particular, the 26th and 27th lines have been changed. On line 26, the word camera has been replaced with miscellaneous. On line 27, the word story has been replaced with miscellaneous.
The second thing to note is that the file src/miscellanous/mod.rs has been added. This file contains a plugin group for miscellaneous plugins.

The third thing to note is that the file src/miscellanous/textbox.rs has been added. This file contains a plugin for a text box.

The fourth thing to note is that the file src/story/new_game.rs has been changed. In particular, the 8th and 23rd lines have been changed. On line 8, the word add has been replaced with remove. On line 23, the word remove has been replaced with add.