The first diff is for the file textbox.rs. In this file, the TextBox plugin is being implemented. The plugin is being built to add a textbox to the app. The textbox is being created with the new() function, which takes in a boolean value and a vector of strings. The boolean value is being set to true, and the vector of strings is being set to an empty string. The textbox is then being added to the app with the add_startup_system() function.
The second diff is for the file new_game.rs. In this file, the NewGame plugin is being implemented. The plugin is being built to add a new game to the app. The game is being created with the new() function, which takes in a boolean value and a vector of strings. The boolean value is being set to true, and the vector of strings is being set to an empty string. The game is then being added to the app with the add_startup_system() function.