The first thing to note is that the diff is for multiple files. In particular, it is for the file src/miscellanous/textbox.rs.
The diff shows that the plugin TextboxPlugin has been modified to add a new system, TextBox::spawn. This system will be responsible for spawning a new textbox on the screen. Additionally, the file has been modified to add a new system, TextBox::detect_change. This system will be responsible for detecting changes to the textbox and updating the display accordingly.