The first diff is for the file src/chat/component.rs. In this file, the new function has been made public. This function is used to create a new chat box sprite.
The second diff is for the file src/chat/controller.rs. In this file, the new function has been made public. This function is used to create a new chat box text.

The third diff is for the file src/chat/resource.rs. In this file, the spawn, dialogue_index, and dialgoues variables have been made public. These variables are used to control the chat box.