The first thing to note is that the name of the MenuComponent struct has been changed to MenuScreenComponent. This is probably to avoid confusion with other components that might be used in the menu screen.
Next, the background() function has been changed to use the new MenuScreenComponent struct. This function is responsible for loading the background image and setting the size.

Finally, the controller.rs file has been updated to use the new MenuScreenComponent struct. This file contains the code that controls the menu screen.